#50: Error esborrant bústia amb continguts
Task-Url: https://github.com/GovernIB/distribucio/issues/50

+++ distribucio-core/src/main/java/es/caib/distribucio/core/service/BustiaServiceImpl.java
@@ -353,6 +353,18 @@ public class BustiaServiceImpl implements BustiaService {
 				}			
 			}
 		}
+		
+		// cannot remove busties containing any anotacions
+		if (contingutRepository.findByPare(bustia) != null && !contingutRepository.findByPare(bustia).isEmpty()) {
+			String missatgeError = "No es pot esborrar la bústia amb anotacions a dins (" + 
+					"bustiaId=" + id + ")";
+			logger.error(missatgeError);
+			throw new ValidationException(
+					id,
+					BustiaEntity.class,
+					missatgeError);
+		}
+		
 		bustiaRepository.delete(bustia);
 		return toBustiaDto(
 				bustia,
